+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/concepts/Concept.kt
@@ -25,12 +25,7 @@
  */
 package de.fraunhofer.aisec.cpg.graph.concepts
 
-import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.OverlayNode
-import de.fraunhofer.aisec.cpg.graph.edges.Edge
-import de.fraunhofer.aisec.cpg.graph.edges.collections.EdgeSingletonList
-import de.fraunhofer.aisec.cpg.graph.edges.collections.MirroredEdgeCollection
-import kotlin.reflect.KProperty
 
 /**
  * Represents a new concept added to the CPG. This is intended for modelling "concepts" like
@@ -41,21 +36,3 @@ abstract class Concept<T : Operation>() : OverlayNode() {
     /** All [Operation]s belonging to this concept. */
     val ops: Set<T> = setOf()
 }
-
-class OverlayEdge(start: Node, end: Node) : Edge<Node>(start, end) {
-    override var labels: Set<String> = setOf("OVERLAY")
-}
-
-class OverlaySingleEdge(
-    thisRef: Node,
-    of: Node?,
-    override var mirrorProperty: KProperty<MutableCollection<OverlayEdge>>,
-    outgoing: Boolean = true,
-) :
-    EdgeSingletonList<Node, Node?, OverlayEdge>(
-        thisRef = thisRef,
-        init = ::OverlayEdge,
-        outgoing = outgoing,
-        of = of,
-    ),
-    MirroredEdgeCollection<Node, OverlayEdge>
