+++ cpg-neo4j/src/test/kotlin/de/fraunhofer/aisec/cpg_vis_neo4j/Neo4JTest.kt
@@ -77,6 +77,7 @@ class Neo4JTest {
 
         val connect = Connect(concept = nw)
         connect.underlyingNode = connectCall
+        connect.name = Name("connect")
 
         assertEquals(connect, connectCall?.overlayNode)
